requirements.txt
streamlit==1.4.0
openai==1.54.3
python-dotenv==1.0.1